Allow explicit control over whether unwind to C is treated as uncaught.
The original Unity hack tries to mark exceptions thrown from MonoBehavior
methods and caught by C->C# bridge as uncaught so that the debugger will
stop on them. This doesn't work with exceptions that are thrown by
libraries like even Unity methods. Also, it seems impossible to disable
stopping on uncaught exceptions in MonoDevelop.

This commit makes it instead consider all exceptions unwinding C uncaught,
provided that the debugger is asked to stop on StackOverflowException
(this exception seems unlikely to happen without causing a total crash)